ocr/requirements.txt
Django>=5.0
djangorestframework
pillow
numpy
opencv-python-headless
paddleocr
